services/notification-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
